+++ src/app/locales/en/components/share-show/header-management.js
@@ -4,4 +4,5 @@ export default {
   publishOpenData: 'Publish as Open Data',
   rename: 'Rename',
   remove: 'Remove',
+  pathForDeletedRootFile: 'Unknown (files deleted)',
 };